refactor package name import
refactor import org.guanzon.cas.model.inventory to import org.guanzon.cas.inventory.models for all inventory validator class
